refactor: refactoring LoginEntranceView
